+++ px/main.py
@@ -1,6 +1,8 @@
 import logging
 import sys
 
+import pkg_resources
+
 from .flatpak import Flatpak
 from .guix import Guix
 from .log import *
@@ -8,12 +10,16 @@ from .messages import help_block
 
 log = logging.getLogger(__name__)
 
+version = pkg_resources.require("px")[0].version
+
 
 def main():
     argument_count = len(sys.argv)
 
     if argument_count < 2:
-        log.error('There is nothing to do here.')
+        log.info('PantherX System Manager v{}'.format(version))
+        print("")
+        help_block()
         sys.exit(0)
 
     if sys.argv[1] == "update":